services/requirements.txt
requests>=2.28.0
websockets>=12.0
aiohttp>=3.9.0
pyarrow>=14.0.0
pandas>=2.0.0
pyyaml>=6.0
cryptography>=41.0.0
numpy>=1.24.0

# NWP models (HRRR, RRFS, NBM, RTMA-RU) — research/download_data/
# GRIB2 data from NOAA cloud archives via Herbie
herbie-data>=2024.1.0
# NBM COG (noaa-nbm-pds)
rasterio>=1.3.0
pyproj>=3.0.0
xarray>=2023.1.0
cfgrib>=0.9.10
eccodes>=1.5.0

# NWP SNS listener — AWS SQS/SNS/S3 for real-time model notifications
boto3>=1.28.0

# MADIS observation ingest — gzipped NetCDF from noaa-madis-pds
netCDF4>=1.6.0

# ── Weather Brain ML Model Pipeline ────────────────────────────────────
# Stage 2: XGBoost quantile regression (reg:quantileerror)
xgboost>=2.0.0
# Stage 5: Isotonic calibration (IsotonicRegression) + evaluation metrics
scikit-learn>=1.3.0
